+++ core/src/test/java/com/transferwise/common/gracefulshutdown/test/StrategyB.java
@@ -1,0 +1,10 @@
+package com.transferwise.common.gracefulshutdown.test;
+
+import javax.annotation.Priority;
+import org.springframework.stereotype.Component;
+
+@Component
+@Priority(1)
+public class StrategyB extends BaseOrderStrategy {
+
+}